added BrandRepo, CategoryRepo, ReviewRepo, BrandDao, CategoryDao and ReviewDao

+++ src/main/java/edu/goshop_ecommerce/repo/BrandRepo.java
@@ -2,8 +2,8 @@ package edu.goshop_ecommerce.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import edu.goshop_ecoomerce.entity.Brand;
+import edu.goshop_ecommerce.entity.Brand;
 
-public interface BrandRepo extends JpaRepository<Brand, Integer> {
+public interface BrandRepo extends JpaRepository<Brand, Long> {
 
 }
