+++ src/main/java/edu/goshop_ecommerce/repo/CategoryRepo.java
@@ -2,8 +2,8 @@ package edu.goshop_ecommerce.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import edu.goshop_ecoomerce.entity.Category;
+import edu.goshop_ecommerce.entity.Category;
 
-public interface CategoryRepo extends JpaRepository<Category, Integer> {
+public interface CategoryRepo extends JpaRepository<Category, Long> {
 
 }
