+++ src/main/java/edu/goshop_ecommerce/repo/ReviewRepo.java
@@ -2,8 +2,8 @@ package edu.goshop_ecommerce.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import edu.goshop_ecoomerce.entity.Review;
+import edu.goshop_ecommerce.entity.Review;
 
-public interface ReviewRepo extends JpaRepository<Review, Integer>{
+public interface ReviewRepo extends JpaRepository<Review, Long> {
 
 }
